Move renderSync tests to only run on Node
