+++ js-api-spec/deprecations.test.ts
@@ -89,30 +89,6 @@ describe('a warning', () => {
       },
     });
   });
-
-  describe('from the JS API', () => {
-    it('is emitted with no flags', done => {
-      renderSync({
-        data: 'a { b: c; }',
-        logger: {
-          warn(message: string) {
-            expect(message).toContain('legacy JS API is deprecated');
-            done();
-          },
-        },
-      });
-    });
-
-    it('is not emitted when deprecation silenced', () => {
-      const stdio = captureStdio(() => {
-        renderSync({
-          data: 'a { b: c; }',
-          silenceDeprecations: [deprecations['legacy-js-api']],
-        });
-      });
-      expect(stdio.err).toBe('');
-    });
-  });
 });
 
 describe('an error', () => {
